Migrate pointerevents/mouse-on-object.html to wpt
The expected behavior on this test seems to be a
hack as it implicitly captures mouse to the object
and still sends the mouseup/pointerup to the object
when down happened on the object even though up is
happening outside of the object. This seems to
resolving compat issues on embedded flash objects.

Change-Id: Ibd888a989dbcf5e6ff08b35569903967edc957e6