Adds slotchange event for Imperative Slot API.
Prior to this CL slotchange event is not fired when calling the
Imperative Slot API.

This CL adds the ability to fire slotchange event for Imperative Slot
API. It does so by detecting candidates nodes changed during the
call and fires slotchange event if changed. This CL only implements the
assign() function's portion of the slotchange event. In the next CL,
I will implement slotchange event for removing nodes and slots.

Bug:869308
Change-Id: I13851d1788fc6cf8b0ea109fb136a2f029e61e70